@@ -12,31 +12,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.11",
-    "@angular/cdk": "^5.2.5",
-    "@angular/common": "^5.2.11",
-    "@angular/compiler": "^5.2.11",
-    "@angular/core": "^5.2.11",
-    "@angular/flex-layout": "^5.0.0-beta.14",
-    "@angular/forms": "^5.2.11",
-    "@angular/http": "^5.2.11",
-    "@angular/material": "^5.2.5",
-    "@angular/platform-browser": "^5.2.11",
-    "@angular/platform-browser-dynamic": "^5.2.11",
-    "@angular/router": "^5.2.11",
+    "@angular/animations": "6.0.2",
+    "@angular/cdk": "^6.0.2",
+    "@angular/common": "6.0.2",
+    "@angular/compiler": "6.0.2",
+    "@angular/core": "6.0.2",
+    "@angular/flex-layout": "^6.0.0-beta.15",
+    "@angular/forms": "6.0.2",
+    "@angular/http": "6.0.2",
+    "@angular/material": "^6.0.2",
+    "@angular/platform-browser": "6.0.2",
+    "@angular/platform-browser-dynamic": "6.0.2",
+    "@angular/router": "6.0.2",
     "@ngx-translate/core": "^9.1.1",
     "angular2-jwt": "^0.2.3",
     "angular2-moment": "^1.9.0",
     "core-js": "^2.5.6",
     "hammerjs": "^2.0.8",
     "ng2-webstorage": "^2.0.0",
-    "rxjs": "^5.5.10",
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "6.1.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.11",
-    "@angular/language-service": "^5.2.11",
+    "@angular-devkit/build-angular": "^0.6.3",
+    "@angular/cli": "^6.0.3",
+    "@angular/compiler-cli": "6.0.2",
+    "@angular/language-service": "6.0.2",
     "@types/jasmine": "^2.8.7",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.111",
@@ -51,6 +53,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "2.7.2"
   }
 }
