@@ -9,13 +9,13 @@ export class SidenavService {
     this.sidenav = sidenav;
   }
 
-  public open(): Promise<void> {
-    return this.sidenav.open();
-  }
+  // public open(): Promise<void> {
+  //   return this.sidenav.open();
+  // }
 
-  public close(): Promise<void> {
-    return this.sidenav.close();
-  }
+  // public close(): Promise<void> {
+  //   return this.sidenav.close();
+  // }
 
   // public toggle(isOpen?: boolean): Promise<void> {
   //   return this.sidenav.toggle(isOpen);
